# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "test oauth API",
        "version": "0.3"
    },
    "paths": {
        "/api/device": {
            "get": {
                "tags": [
                    "Device"
                ],
                "summary": "Get Device",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "include",
                        "in": "query",
                        "description": "relationships to include (Comma seperated)",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "filter[id]",
                        "in": "query",
                        "description": "id of device",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "filter[ip]",
                        "in": "query",
                        "description": "ip of device",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "filter[type]",
                        "in": "query",
                        "description": "type of device",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "filter[name]",
                        "in": "query",
                        "description": "name of device",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "filter[model]",
                        "in": "query",
                        "description": "model of device",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "filter[serial]",
                        "in": "query",
                        "description": "serial of device",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "AzureAD": []
                    }
                ]
            }
        },
        "/api/device/{id}": {
            "get": {
                "tags": [
                    "Device"
                ],
                "summary": "Get Device by ID",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Device",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                },
                "security": [
                    {
                        "AzureAD": []
                    }
                ]
            }
        },
        "/api/device/search/{search}": {
            "get": {
                "tags": [
                    "Device"
                ],
                "summary": "Get Device by Search",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "search",
                        "in": "path",
                        "description": "search",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                },
                "security": [
                    {
                        "AzureAD": []
                    }
                ]
            }
        },
        "/api/hello": {
            "get": {
                "summary": "Hello world test for API troubleshooting",
                "responses": {
                    "200": {
                        "description": "Hello world example"
                    }
                }
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "Me"
                ],
                "summary": "Get user information about the token authenticated user",
                "responses": {
                    "200": {
                        "description": "User information"
                    },
                    "401": {
                        "description": "Authentication failed"
                    }
                },
                "security": [
                    {
                        "AzureAD": []
                    }
                ]
            }
        },
        "/api/me/roles": {
            "get": {
                "tags": [
                    "Me"
                ],
                "summary": "Get user roles for the token authenticated user",
                "responses": {
                    "200": {
                        "description": "User information"
                    },
                    "401": {
                        "description": "Authentication failed"
                    }
                },
                "security": [
                    {
                        "AzureAD": []
                    }
                ]
            }
        },
        "/api/me/roles/permissions": {
            "get": {
                "tags": [
                    "Me"
                ],
                "summary": "Get user role permissions for the token authenticated user",
                "responses": {
                    "200": {
                        "description": "User information"
                    },
                    "401": {
                        "description": "Authentication failed"
                    }
                },
                "security": [
                    {
                        "AzureAD": []
                    }
                ]
            }
        }
    },
    "definitions": {},
    "securityDefinitions": {
        "AzureAD": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://login.microsoftonline.com/07420c3d-c141-4c67-b6f3-f448e5adb67b/oauth2/v2.0/authorize",
            "scopes": {
                "api://915c46fe-ee91-41c7-98ab-b257b04ea7ec/access_as_user": "Use client_id: 915c46fe-ee91-41c7-98ab-b257b04ea7ec"
            }
        }
    }
}